+++ Puzzles/Days/Day1/PuzzleDay1a.cs
@@ -3,6 +3,6 @@
     public class PuzzleDay1a : PuzzleDay1
     {
         protected override int numberOfComponents { get => 2; }
-        protected override int sumTo { get => 2020; }
+        protected override ulong sumTo { get => 2020; }
     }
 }
