+++ Puzzles/Days/Day1/PuzzleDay1b.cs
@@ -3,7 +3,7 @@
     public class PuzzleDay1b : PuzzleDay1
     {
         protected override int numberOfComponents { get => 3; }
-        protected override int sumTo { get => 2020; }
+        protected override ulong sumTo { get => 2020; }
 
     }
 }
